when restoring the sampler, try to update the queue_size of the sampler when using the pool
